Check and reset `is_set` property during parsing
Previously it was possible for arguments to become unset during parsing
due each argument object being evaluated for each argument string in the
command line. This change adds checks to each argument affected by the
issue for whether the argument is already set. In order to allow parsers
to be re-used, the `is_set` property of each argument is reset at the
beginning of each command line parsing.